✍️ Refactor `televerse.dart` to `bot.dart`
