+++ lib/src/televerse/bot.dart
