+++ lib/televerse.dart
@@ -35,7 +35,7 @@ export 'src/televerse/fetch/fetch.dart';
 export 'src/televerse/extensions/extensions.dart';
 export 'src/televerse/links/links.dart';
 
-part 'src/televerse/televerse.dart';
+part 'src/televerse/bot.dart';
 part 'src/utils/date.dart';
 part 'src/utils/utils.dart';
 part 'src/televerse/markups/keyboard.dart';
